chore: projen touchups (#5)
* mid work

* bump to node 16 and other touchups

* bump to 16.14.0
