+++ .github/workflows/build.yml
@@ -25,7 +25,7 @@ jobs:
       - name: Setup Node.js
         uses: actions/setup-node@v3
         with:
-          node-version: 16.x
+          node-version: 16.14.0
       - name: Install dependencies
         run: yarn install --check-files
       - name: build
@@ -94,7 +94,7 @@ jobs:
     steps:
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.x
+          node-version: 16.14.0
       - name: Download build artifacts
         uses: actions/download-artifact@v3
         with:
@@ -123,7 +123,7 @@ jobs:
           java-version: 11.x
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.x
+          node-version: 16.14.0
       - name: Download build artifacts
         uses: actions/download-artifact@v3
         with:
@@ -148,7 +148,7 @@ jobs:
     steps:
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.x
+          node-version: 16.14.0
       - uses: actions/setup-python@v4
         with:
           python-version: 3.x
@@ -176,7 +176,7 @@ jobs:
     steps:
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.x
+          node-version: 16.14.0
       - uses: actions/setup-dotnet@v3
         with:
           dotnet-version: 3.x
@@ -204,7 +204,7 @@ jobs:
     steps:
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.x
+          node-version: 16.14.0
       - uses: actions/setup-go@v3
         with:
           go-version: ^1.16.0
